# packages/oas/specification/v1.yml
openapi: 3.0.3
info:
  version: 0.0.1
  title: Decentralized-Guild
  description: About Decentralized Guild System APIs. Please use it when expanding the system.
  termsOfService: 'https://github.com/ymuichiro/decentralized-guild/blob/main/LICENSE'
  contact:
    email: ym.u.ichiro@gmail.com
    name: E-mail
    url: 'https://github.com/ymuichiro/decentralized-guild'
  license:
    name: MIT
    url: 'https://github.com/ymuichiro/decentralized-guild/blob/main/LICENSE'
...
servers:
  - url: '/api'
...
paths:
...
  /users/{userId}:
    get:
      tags: [user]
      summary: get user info
      operationId: GetUser
      description: ユーザー情報を取得する。SessionID と同一の場合のみ email 情報を返す。
      parameters: [$ref: '#/components/parameters/UserIdPath']
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [user]
      summary: put user info
      description: ユーザーの情報を更新する。パブリックキー、プロフィールの更新は別APIにて定義する
      operationId: PutUser
      parameters: [$ref: '#/components/parameters/UserIdPath']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: PutUserRequestBody
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '204':
          description: SUCCESS NO CONTENT
    delete:
      tags: [user]
      summary: delete current user
      operationId: DeleteUser
      description: ユーザーの退会処理。パブリックキー情報はサーバー上に残置される。
      parameters: [$ref: '#/components/parameters/UserIdPath']
      responses:
        '204':
          description: SUCCESS NO CONTENT
...
  /quests/{questId}/workers:
    get:
      tags: [quest]
      summary: get quest workers
      operationId: GetQuestWorkers
      description: クエストに紐づくワーカーの一覧を取得する
      parameters: [$ref: '#/components/parameters/QuestIdPath']
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                type: object
                title: GetQuestWorkersResponse
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/QuestWorker'
    post:
      tags: [quest]
      summary: post quest worker
      operationId: PostQuestWorker
      description: クエストにワーカーを追加する
      parameters: [$ref: '#/components/parameters/QuestIdPath']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: PostQuestWorkerRequestBody
              required: [proposal]
              properties:
                proposal:
                  type: string
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatRoom'
...
  /chats:
    get:
      tags: [chat]
      summary: get chat rooms
      operationId: GetChats
      description: 自身がアクセス可能なチャットルームの一覧を取得
      parameters: [$ref: '#/components/parameters/OffsetPath']
      responses:
        '200':
          $ref: '#/components/responses/GetChatRoomsResponse'
    post:
      tags: [chat]
      summary: post chat room
      operationId: PostChat
      description: post chat room
      requestBody:
        $ref: '#/components/requestBodies/PostChatRoomRequest'
      responses:
        '200':
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatRoom'
    put:
      tags: [chat]
      summary: put all chat room is read
      operationId: putAllChatRoomIsRead
      description: put all chat room is read
      responses:
        '204':
          description: SUCCESS
  /chats/{chatRoomId}:
    get:
      tags: [chat]
      summary: get chat room info
      operationId: GetChatInfo
      description: get chat room info
      parameters: [$ref: '#/components/parameters/ChatRoomIdPath']
      responses:
        '200':
          description: SUCCESSFUL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatRoom'
    put:
      tags: [chat]
      summary: is read chat room
      operationId: PutChatInfo
      description: is read chat room
      parameters: [$ref: '#/components/parameters/ChatRoomIdPath']
      responses:
        '204':
          description: SUCCESSFUL
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            user: Read user info
  parameters:
    ChatRoomIdPath:
      name: chatRoomId
      in: path
      required: true
      schema:
        type: string
...
    UserIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
    QuestIdPath:
      name: questId
      in: path
      required: true
      schema:
        type: string
...
    OffsetPath:
      name: offset
      in: query
      description: Number of items to skip before returning the results.
      required: false
      schema:
        type: number
        minimum: 0
        default: 0
  responses:
...
    GetChatRoomsResponse:
      description: SUCCESSFUL
      content:
        application/json:
          schema:
            title: GetChatRoomsResponseInner
            type: object
            required: [data, pageNation]
            properties:
              pageNation:
                $ref: '#/components/schemas/PageNation'
              data:
                type: array
                items:
                  $ref: '#/components/schemas/ChatRoom'
...
  requestBodies:
...
    PostChatRoomRequest:
      required: true
      description: post chat room request
      content:
        application/json:
          schema:
            type: object
            title: PostChatRoomRequestInner
            required: [data]
            properties:
              data:
                type: object
                title: PostChatRoomRequestInnerData
                required: [recipients]
                properties:
                  recipients:
                    type: array
                    items:
                      type: string
  schemas:
    PageNation:
      type: object
      required:
        - id
        - total
        - current
        - last
      properties:
        total:
          type: number
        current:
          type: number
        last:
          type: number
    User:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
          description: base64 encoding
        publicKey:
          type: string
          description: Symbol public key
        createdAt:
          type: string
          description: UTC. If you have already canceled your membership, only publickey will remain, so it will be undefined.
          format: date-time
        email:
          type: string
          description: user email
...
    QuestWorker:
      type: object
      required: [id, questId, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        questId:
          type: string
        status:
          $ref: '#/components/schemas/QuestWorkerStatusEnum'
        worker:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
    QuestWorkerStatusEnum:
      type: string
      enum: [APPLYING, ACCEPTED, COMPLETED, REJECTED]
...
    ChatRoom:
      type: object
      required:
        - id
        - createdAt
        - lastPostedUserId
        - updatedAt
        - messages
        - members
        - isUnread
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        lastPostedUserId:
          type: string
        updatedAt:
          type: string
          format: date-time
        isUnread:
          type: boolean
          description: unread message exists current user
        messages:
          type: array
          items:
            type: object
            title: ChatRoomMessage
            required:
              - id
              - body
              - createdAt
            properties:
              id:
                type: string
              body:
                type: string
              createdAt:
                type: string
                format: date-time
              userId:
                type: string
        members:
          type: array
          items:
            type: object
            title: ChatRoomMember
            required: [id]
            properties:
              id:
                type: string
              userId:
                type: string
              name:
                type: string
...